gittf/requirements.txt
requests
pyyaml
aws-lambda-powertools
boto3
click
pyyaml
awscli
PyJWT